I'm Martian

+++ src/MINEOStoDSMmodels/modules.f90
@@ -18,7 +18,7 @@ module parameters
   character(120) :: psvmodel
   integer :: nzone
   double precision, allocatable, dimension (:) :: vrmin, vrmax,qmuD,qkappaD
-  double precision, allocatable, dimension (:) :: irmin, irmax,irlength
+  integer, allocatable, dimension (:) :: irmin, irmax,irlength
   double precision, allocatable, dimension (:,:) :: rrhoD,vpvD,vphD,vsvD,vshD,etaD
